small UI
I clean up the logo to remove the mb from the div that was creating a lot of space between the top. Also updates the img on the profile page to show the same profile page as the appHeader. Also IF the provider is git i remove the updload image since that is coming from github.